+++ mcp_server/src/codebase_indexer/colbert_worker/__init__.py
@@ -1,1 +1,0 @@
-"""ColBERT HTTP sidecar package."""